% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PghTracts.R
\docType{data}
\name{PghTracts}
\alias{PghTracts}
\title{Drug Offenses}
\format{
A data frame with 144 rows and 4 columns containing the census tract and the variables \code{YEAR},\code{MONTH} and \code{DRUGS}.
}
\source{
Aldo M.; Medina, Francyelle L.; Jales, Isaac C.; Bertail, Patrice. First-order integer valued AR processes with zero-inflated innovations. Cyclostationarity: Theory and Methods, Springer Verlag - 2021, v. 1, p. 19-40. DOI: 10.1007/978-3-030-82110-4_2
}
\usage{
PghTracts
}
\description{
Monthly number of drug offenses recorded from January 1990 to December 2001, with 144 observations, in Pittsburgh census tract 2206.
}
\keyword{datasets}
